build: Add a toggle to override the module maps
This adds a toggle in `build.ps1`, off by default, to override the
module maps in the bootstrap toolchain installation folder so they work
with the latest version of the Windows SDK. The toggle is explicitly
marked as a hack and disabled by default to discourage use. The log
information also makes it clear that a hack has been enabled.